requirements.txt
python-frontmatter>=3.0.0
SQLAlchemy>=2.0.0
FastMCP==0.4.1